Refactor scoring and move choice in MinMaxPlayer class
The scoring method was moved from the game board to the MinMaxPlayer class. This approach allows each player to deduct their own scores, enhancing modularity and code maintainability. The depth parameter was also replaced with ply, with ply / 2 being the new depth, making the scoring closer to the way it is in actual game play. Additionally, the random.shuffle method in the move choice method was commented out to adhere to a consistent move decision strategy and facilitate debugging, and the pressing of enter for progressing to the next move was made optional for convenience during debugging.